Update practitionerDirectory.fsh
Not clear to me that we have the right AND / OR logic documented yet. The example in this PR is a modification of an example already in the IG. The example is intended to enable HR and IT to make full changes to resources in the directory. It is not clear to me this is done with a repetition of purpose element, or if one must repeat at the activity level.